+++ litellm/integrations/gcs_pubsub/pub_sub.py
@@ -41,6 +41,10 @@ class GcsPubSubLogger(CustomBatchLogger):
             topic_id (str): Pub/Sub topic ID
             credentials_path (str, optional): Path to Google Cloud credentials JSON file
         """
+        from litellm.proxy.utils import _premium_user_check
+
+        _premium_user_check()
+
         self.async_httpx_client = get_async_httpx_client(
             llm_provider=httpxSpecialProvider.LoggingCallback
         )
@@ -102,6 +106,9 @@ class GcsPubSubLogger(CustomBatchLogger):
         from litellm.proxy.spend_tracking.spend_tracking_utils import (
             get_logging_payload,
         )
+        from litellm.proxy.utils import _premium_user_check
+
+        _premium_user_check()
 
         try:
             verbose_logger.debug(
